+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Bubble chart layer and example, example includes functionality to map model values to colors from gradient bar
 - Scatter chart layer and example
 - Bars layer, +/- bars with dynamic gradient example 
+- Grouped bars layer and example
 
 ### Removed
 - Bars example with horizontal axis - now included in variable axes example